+++ tests/nightly/single_card/sentence_bert/test_ttnn_sentencebert_intermediate.py
@@ -1,0 +1,1 @@
+../../../../models/demos/sentence_bert/tests/pcc/test_ttnn_sentencebert_intermediate.py
